+++ src/entry/editor.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+import { render } from 'react-dom';
+import App from '../lib/app';
+
+render(<App />, document.getElementById('react-root'));