+++ src/entry/preview.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+import { render } from 'react-dom';
+import Preview from '../lib/preview';
+
+render(<Preview />, document.getElementById('mw-root'));