+++ src/lib/preview.tsx
@@ -1,14 +1,25 @@
 import { ipcRenderer as ipc } from 'electron';
-import Markwright from 'markwright';
-import * as React from 'react';
 import * as cache from 'js-cache';
-
+import Markwright from 'markwright';
+import { homedir } from 'os';
+import * as sizes from 'paper-size';
+import * as React from 'react';
 import { proxy } from 'workly';
 
-import * as markwright from './themes/markwright';
+import * as markwright from '../themes/markwright';
 
-export default class Preview extends React.Component<any, any> {
-  public highlight = proxy('../assets/highlight.js');
+type PreviewProps = {};
+type PreviewState = {
+  metadata: any;
+  markdown: string;
+  css: string;
+};
+
+export default class Preview extends React.Component<
+  PreviewProps,
+  PreviewState
+> {
+  public highlight = proxy('../lib/highlight.js');
 
   public mouse = { x: 0, y: 0 };
 
@@ -21,14 +32,15 @@ export default class Preview extends React.Component<any, any> {
     return key in this.state.metadata ? this.state.metadata[key] : value;
   }
 
-  public state = {
+  public state: PreviewState = {
     metadata: {},
     markdown: '',
     css: ''
   };
 
   public get themeCSS() {
-    return markwright.styles(this.state.metadata);
+    const page = this.page;
+    return markwright.styles(this.state.metadata, page.width, page.height);
   }
 
   public get hintCSS() {
@@ -57,7 +69,6 @@ export default class Preview extends React.Component<any, any> {
   }
 
   protected registerToolTipEvents() {
-
     document.addEventListener('mousemove', e => {
       this.mouse.x = e.clientX;
       this.mouse.y = e.clientY;
@@ -102,21 +113,56 @@ export default class Preview extends React.Component<any, any> {
     ipc.on('editor.metadata', (e, metadata) => this.setState({ metadata }));
     ipc.on('editor.markdown', (e, markdown) => this.setState({ markdown }));
     ipc.on('editor.css', (e, css) => this.setState({ css }));
+    ipc.on('editor.base', (e, dir) => {
+      document.head.querySelector('base').setAttribute('href', dir);
+    });
+  }
+
+  public get page() {
+    const paper = this.state.metadata.paper || 'letter';
+    const isPortrait = this.state.metadata.orientation !== 'landscape';
+    const dimensions = sizes.getSize(paper, { unit: 'inches' }) || [216, 279];
+    const [width, height] = isPortrait ? dimensions : dimensions.reverse();
+    return {
+      width: Math.round(width * 20) / 20,
+      height: Math.round(height * 20) / 20
+    };
   }
 
   public componentDidMount() {
-    this.registerIPCEvents();
     this.registerToolTipEvents();
+    this.registerIPCEvents();
+    sizes.register('letter', 216, 279);
+    sizes.register('legal', 216, 356);
+    sizes.register('tabloid', 279, 432);
+    sizes.register('half-letter', 140, 216);
+    document.head.querySelector('base').setAttribute('href', homedir() + '/');
+
+    ipc.send('app.ready');
   }
 
   public render() {
+    const page = this.page;
     return (
-      <div>
-        <style type="text/css">{this.state.css}</style>
+      <main
+        style={{
+          paddingTop: "2.5rem",
+          height: '100%',
+          width: '100%',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center'
+        }}
+      >
         <style type="text/css">{this.themeCSS}</style>
+        <style type="text/css">{this.state.css}</style>
         <style type="text/css">{this.hintCSS}</style>
         <Markwright
           config={{
+            page: {
+              width: page.width * 96,
+              height: page.height * 96
+            },
             manual: this.metadata('manual', false),
             highlight: (str, lang) => {
               const key = `${lang}::${str}`;
@@ -131,8 +177,9 @@ export default class Preview extends React.Component<any, any> {
             }
           }}
           value={this.state.markdown}
+          page={1}
         />
-      </div>
+      </main>
     );
   }
 }